Add 1.18.2 support

+++ src/me/bomb/cutscene/CameraManager_v1_13_R2.java
@@ -61,7 +61,7 @@ private static final PacketPlayOutWindowItems packetemptywindowitems;
 	
 	static {
 		NonNullList<ItemStack> nnl = NonNullList.a();
-		for (int slot = 0; slot < 46; slot++) {
+		for (byte slot = 0; slot < 46; slot++) {
 			nnl.add(slot, new ItemStack(Item.getById(0)));
 		}
 		packetemptywindowitems = new PacketPlayOutWindowItems(0, nnl);
